Consolodate errors
